refactor search paging
