+++ sdk/search/azure-search/azure/search/_index/aio/_search_index_client_async.py
@@ -7,54 +7,18 @@ from typing import cast, List, TYPE_CHECKING
 
 import six
 
-from azure.core.async_paging import AsyncItemPaged, AsyncPageIterator
 from azure.core.pipeline.policies import HeadersPolicy
 from azure.core.tracing.decorator_async import distributed_trace_async
+from ._paging import AsyncSearchItemPaged, AsyncSearchPageIterator
 from .._generated.aio import SearchIndexClient as _SearchIndexClient
 from .._generated.models import IndexBatch, IndexingResult, SearchRequest
 from .._index_documents_batch import IndexDocumentsBatch
 from .._queries import AutocompleteQuery, SearchQuery, SuggestQuery
-from .._search_index_client import (
-    convert_search_result,
-    pack_continuation_token,
-    unpack_continuation_token,
-)
 
 if TYPE_CHECKING:
     # pylint:disable=unused-import,ungrouped-imports
     from typing import Any, Union
     from .._credential import SearchApiKeyCredential
-
-
-class _SearchDocumentsPagedAsync(AsyncPageIterator):
-    def __init__(self, client, initial_query, kwargs, continuation_token=None):
-        super(_SearchDocumentsPagedAsync, self).__init__(
-            get_next=self._get_next_cb,
-            extract_data=self._extract_data_cb,
-            continuation_token=continuation_token,
-        )
-        self._client = client
-        self._initial_query = initial_query
-        self._kwargs = kwargs
-
-    async def _get_next_cb(self, continuation_token):
-        if continuation_token is None:
-            return await self._client.documents.search_post(
-                search_request=self._initial_query.request, **self._kwargs
-            )
-
-        _next_link, next_page_request = unpack_continuation_token(continuation_token)
-
-        return await self._client.documents.search_post(
-            search_request=next_page_request
-        )
-
-    async def _extract_data_cb(self, response):  # pylint:disable=no-self-use
-        continuation_token = pack_continuation_token(response)
-
-        results = [convert_search_result(r) for r in response.results]
-
-        return continuation_token, results
 
 
 class SearchIndexClient(object):
@@ -139,7 +103,7 @@ class SearchIndexClient(object):
 
     @distributed_trace_async
     async def search(self, query, **kwargs):
-        # type: (Union[str, SearchQuery], **Any) -> AsyncItemPaged[dict]
+        # type: (Union[str, SearchQuery], **Any) -> AsyncSearchItemPaged[dict]
         """Search the Azure search index for documents.
 
         :param query: An query for searching the index
@@ -173,8 +137,8 @@ class SearchIndexClient(object):
                 )
             )
 
-        return AsyncItemPaged(
-            self._client, query, kwargs, page_iterator_class=_SearchDocumentsPagedAsync
+        return AsyncSearchItemPaged(
+            self._client, query, kwargs, page_iterator_class=AsyncSearchPageIterator
         )
 
     @distributed_trace_async
